+++ creme/feature_selection/random.py
@@ -18,18 +18,15 @@ class RandomDiscarder(base.Transformer):
             >>> discarder = feature_selection.RandomDiscarder(n_to_keep=5)
 
             >>> for x, _ in stream.iter_numpy(X):
-            ...     x_pruned = discarder.fit_one(x)
+            ...     x_pruned = discarder.transform_one(x)
             ...     assert len(x_pruned) == 5
-            ...     # creme's transformers are pure so the input should not have changed
+            ...     # creme's transformers are pure, so the input should not change
             ...     assert len(x) == 8
 
     """
 
-    def __init__(self, n_to_keep=None):
+    def __init__(self, n_to_keep):
         self.n_to_keep = n_to_keep
-
-    def fit_one(self, x, y=None):
-        return self.transform_one(x)
 
     def transform_one(self, x):