+++ creme/utils/__init__.py
@@ -7,7 +7,6 @@ from .math import norm
 from .math import prod
 from .math import sigmoid
 from .math import softmax
-from .nlp import VectorizerMixin
 from .sdft import SDFT
 from .skyline import Skyline
 from .window import Window
@@ -27,6 +26,5 @@ __all__ = [
     'Skyline',
     'softmax',
     'SortedWindow',
-    'VectorizerMixin',
     'Window'
 ]