+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed a bug where `utils.dot` could take longer than necessary
 - Cythonized `stats.Mean` and `stats.Var`
 - Cythonized all the loss functions in `optim`
+- `CountVectorizer` and `TFIDFVectorizer` won't raise an error when being pickled anymore
+- Tests are now much more extensive, thanks mostly to estimator tags
 
 ## [0.3.0](https://pypi.org/project/creme/0.3.0/) - 2019-06-23